+++ jdaviz/configs/default/plugins/metadata_viewer/metadata_viewer.py
@@ -23,7 +23,7 @@ class MetadataViewer(PluginTemplateMixin, DatasetSelectMixin):
       Dataset to expose the metadata.
     * :attr:`show_primary`:
       Whether to show MEF primary header metadata instead.
-    * ``meta``:
+    * :attr:`meta`:
       Read-only metadata. If the data is loaded from a multi-extension FITS file,
       this can be the extension header or the primary header, depending on
       ``show_primary`` setting.
@@ -39,7 +39,6 @@ class MetadataViewer(PluginTemplateMixin, DatasetSelectMixin):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.meta = {}
         # override the default filters on dataset entries to require metadata in entries
         self.dataset.add_filter('not_from_plugin')
 
@@ -48,7 +47,8 @@ class MetadataViewer(PluginTemplateMixin, DatasetSelectMixin):
 
     @property
     def user_api(self):
-        return PluginUserApi(self, expose=('dataset', 'show_primary'), readonly=('meta',))
+        return PluginUserApi(self, expose=('dataset', 'show_primary'),
+                             readonly=('metadata', 'meta'), deprecated=('metadata', ))
 
     def reset(self):
         self.has_metadata = False
@@ -56,7 +56,6 @@ class MetadataViewer(PluginTemplateMixin, DatasetSelectMixin):
         self.show_primary = False
         self.has_comments = False
         self.metadata = []
-        self.meta = {}
 
     @observe("dataset_selected")
     def show_metadata(self, event):
@@ -102,7 +101,8 @@ class MetadataViewer(PluginTemplateMixin, DatasetSelectMixin):
 
         d = flatten_nested_dict(meta)
         # Some FITS keywords cause "# ipykernel cannot clean for JSON" messages.
-        # Also, we want to hide internal metadata that starts with underscore.
+        # Also, we want to hide internal metadata that starts with underscore
+        # and some "original_" entries from specutils.
         badkeys = (['COMMENT', 'HISTORY', ''] + [k for k in d if k.startswith('_')]
                    + [k for k in d if k.startswith('original_')])
         for badkey in badkeys:
@@ -128,11 +128,15 @@ class MetadataViewer(PluginTemplateMixin, DatasetSelectMixin):
         public_meta = sorted(zip(d.keys(), map(str, d.values()), map(get_comment, d.keys())))
         if len(public_meta) > 0:
             self.metadata = public_meta
-            self.meta = d
             self.has_metadata = True
             self.has_comments = has_comments
         else:
             self.reset()
+
+    @property
+    def meta(self):
+        return dict((key, {"value": val, "description": cmt})
+                    for key, val, cmt in self.metadata)
 
 
 # TODO: If this generalized in stdatamodels in the future, replace with native function.